html/semantics/embedded-content-0/the-img-element/img.complete.html

<!DOCTYPE HTML>
<html>
<!-- Submitted from TestTWF Paris -->
<head>
    <title>DOM img complete Test</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
    
    <link rel="author" title="Anselm Hannemann" href="http://anselm-hannemann.com/" />
    <link rel="help" href="http://www.w3.org/TR/html5/embedded-content-0.html#dom-img-complete" />

    <meta name="assert" content="Tests the complete status of the img-element">

    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

</head>
<body>
    <h1>DOM img.complete Test</h1>

    <img id="imgTestTag">

    <img src="" id="imgTestTag2">

    <img id="imgTestTag3" style="width: 80px; height:auto;">

    <img id="imgTestTag4">

    <script>
        var imageInstance = document.createElement('img');
        imageInstance.style.display = 'none';

        document.body.appendChild(imageInstance);
    </script>

<div id="log"></div>
<script>
test(function () {
    assert_true(document.getElementById("imgTestTag").complete);
}, "img src omitted");

test(function () {
    assert_true(document.getElementById("imgTestTag2").complete);
}, "img src empty");

// test if set to true after img is completely available     
var t = async_test("async src complete test");

t.step(
    function(){
        assert_equals("", document.getElementById("imgTestTag3").src);

        //test if img.complete is set to false if src is changed
        document.getElementById("imgTestTag3").src = '3.jpg';
        test (function () {
            assert_false(document.getElementById("imgTestTag3").complete)
        }, "src changed, should be set to false");
    }
);

document.getElementById("imgTestTag3").onload = t.step_func(
    function(){
        assert_true(document.getElementById("imgTestTag3").complete);
        t.done();
    }
);

// test: The final task that is queued by the networking task source once the resource has been fetched has been queued, but has not yet been run, and the img element is not in the broken state
async_test(function(t) {
    imageInstance.src = 'image-1.jar';
    if(imageInstance.complete == false) {       
        imageInstance.src = 'image-1.jpg';      

        imageInstance.onload = t.step_func(function() {
            
            assert_true(imageInstance.complete, 'Image resource has been fetched but not run yet & image is not in the broken state');

            t.done();           
        }); 

    }
},
'IDL attribute complete returns true when image resource has been fetched but not run yet & image is not in broken state');

// test if broken img does not pass
var t2 = async_test("async src broken test");

t2.step(
    function(){
        document.getElementById("imgTestTag4").src = 'brokenimg.ogg';
        //test if img.complete is set to false if src is changed
        assert_false(document.getElementById("imgTestTag4").complete, "src changed to broken img, should be set to false");
    }
);

var img4 = document.getElementById("imgTestTag4");
img4.onload = img4.onerror = t2.step_func(
    function(event){
        assert_equals(event.type, "error");
        assert_true(document.getElementById("imgTestTag4").complete);
        t2.done();
    }
);
</script>
</body>
</html>
